Add support for multipart/form-data conversions
Upload-based conversions can now be carried out by
calling a POST request on `/convert` using the `file`
parameter.

Example:

	curl -F file=@index.html localhost:8080/convert

The output is a binary stream (much like the GET
endpoint).

This change significantly refactors the way a conversion
target / source is handled. When running a conversion,
we now provide an explicit conversion source. Before, we
relied on an implicit source (that we simply assumed a
Converter will handle).

The conversion source also contains the MIME type of
our conversion target. We can extend this to handle a
range of different conversions in the future.

The old `util` package, and its methods have been
integrated into `source.go`. All URLs are checked w/
cookie support.

A new CloudConvert method was also introduced for quick
upload-based conversions. In the future, it might be
worth using the `inline` download method only, and rely
on our own S3 uploader (that way, we won't be passing
our S3 credentials around).

TODO: refactor middleware, and application-wide config.