Suppress download popup window / menu (#36)
The rationale behind this is that we will never be able to 'run'
a PDF conversion in this scenario (i.e. when a download dialog
popup shows up). We are better off failing / exiting quickly.

This scenario should not occur in most cases. If it does, the
user should download the file for local conversion (e.g. via
file or stdin).

Additionally, `octet-stream` files are handled well when used
with Weaver.