Add support for custom file extensions in a local conversion strategy
Previously, local conversions (i.e. save, and load a conversion
file through `/tmp`) are 'extensionless'.

This works fine for majority of use-cases (HTML to PDF).
However, Chromium provides support for loading MHTML files
(message/rfc822) as long as the extension of the file ends with
`.mhtml`, otherwise it will attempt to download the file or
otherwise load it without the necessary formatting.

This commit allows custom file extensions to be specified using
the `ext` parameter (e.g. `?ext=mhtml`).

Alternatively, we could infer the type of the file through a type
hint provided by the request or from the contents of the file.
But, this will add MHTML specific logic that will pollute the
source code, and deviate the project from its original goal of
converting HTML to PDF. The solution described above allows for
more generic additions in the future.