Fix `media` plugin / hack for style elements
The `media` attribute is automatically removed for stylesheet
links if there are no print styles defined.

Previously however, the aforementioned 'autoremoving' behaviour
only applied to elements with `rel=stylesheet`, and it did not
apply to `style` elements.

This commit addresses the problem by including `style` elements
in the query selector.

---

Fixes #61 (Github)

Signed-off-by: Ian Lai <ian@fyianlai.com>