[Version 3] Add protobuf definitions
Version 3 will be driven by protocol buffers, and a library-first
design. The goal is to build around shared definitions that both
the service layer, and the library layer can (re)use. This will
simplify API design, and hasten the development overall.

New API clients for any languages can be generated easily.
And they can immediately be used to interact with the service.
The only potential problem is that we _cannot_ (or rather, we should
not) encode large messages (e.g. binary files). Additionally,
we are unlikely to rely on (G)RPC in the final design as HTTP
is a more common, and understood construct for end-users.
Large files can also be _converted_ via a multipart form.
Nevertheless, these definitions makes it easier for users to
construct a request / process payload (the API is self-describing
insofar as the available parameters are defined by the generated
protobuf code).