[Version 3] Upgrade athenapdf CLI to use Chrome Remote Debugging Protocol (DevTools)
The CLI converter has been broken into two separate parts:
1. a `runner` library, and 2. an executable CLI entrypoint / program.
This is for greater separation of concerns, and to allow for
greater reusability. For example, the service can now call the
`runner` directly, and it does not need to run conversions against
the compiled CLI program. Though, if it wants to, it is also possible
in the future. It will also be a lot simpler to test.

The converter now relies on Chrome Remote Debugging Protocol (DevTools).
This allows us to access all the following APIs: https://chromedevtools.github.io/devtools-protocol/
Additionally, we are no longer tied to Electron's development,
and release cycle. We can also use a completely headless instance
of Chromium (no `xvfb`). Taken together, it provides us greater
flexibility, and performance enhancements (subjective).

JavaScript plugins will still be possible, and they are packed into
the final binary via `go-bindata`.

TODO: documentation (but, this will probably be consolidated into
a single `README.md`).