[Version 3] Move `weaver` to `cmd/weaver` in preparation for its deprecation
The `weaver` microservice will be rearchitected from the ground up.
It will no longer be using a micro-HTTP framework like Gin.

Docker-related files are not being copied over as the project
layout, and development workflow will be changed as well.

`weaver` is only being kept for reference. It will be removed
before version 3 is released, and when the new service
program is completed.