[Version 3] Refactoring: high-level wrappers for fetching, converting, and uploading
The packages can be used in unsafe / insecure ways.

For example, a fetcher does not check if the target is fetchable beforehand.
Likewise, a converter does not check if a request is convertable, etc.
Thus, it is ultimately up to the caller to perform these checks.

We could bundle the checks in the core interface methods, but that would:

1. put too much responsibility in them;
2. and, it may lead to redundancy / checks being executed multiple times.

The latter may occur if we for example, decide to get, and use a converter
that supports a mime type of X. In such a case, we are already confident
that the converter supports X, and we do not need checks to be executed again.

Furthermore, we would simply assume that all implementations of the core
interface methods are _perfect_. That is, there is no _machine_ rule to
state that checks are required. This is potentially _more dangerous_.

Originally, an interface method of `IsFetchable(..)` / `IsConvertable(..)`
was considered, but the implementations for these methods are likely to
be the same, and thus, it will simply result in redundant, and difficult
to maintain code. It may also obscure the reason something is not
_fetchable_ or _convertable_.

Thus, the solution involves creating high-level wrappers for
fetching, converting, and uploading. They will essentially act
as gatekeepers. This is particularly useful in the case of the
`converter` package as it can be used to perform mime type
validation, and to automatically get a converter if none is specified.
Though, the latter case may be moved to the `process` package in
the future for better separation of concerns, and consistency.

There are other security considerations that will have to be
accounted for by the caller.

For example, we probably do not want to allow local conversions
via user input.