[Version 3] Move logic for converter autogetting from `converter` to `process`
As described in the previous commit, the high-level wrapper
methods should not be responsible for such logic. They should simply
act as gatekeepers, and not perform any _unexpected magic_.
That is the responsibility of the process (yet another higher-level
abstraction).

This change also introduces types for various curried functions.
In doing so, we can cut redundancy.

For example:

```go
if converterName == "" {
    c, err := converter.GetFromConversionExcluding(conversion)([]string{})
    if err != nil {
        return nil, false, err
    }
    converterFn = c.Convert
} else {
    converterFn = converter.Convert(converterName)
}
```

The code above allows us to specify the arguments once using:

```go
converterFn(ctx, p.GetConversion(), p.GetConverterOptions())
```