[Version 3] Fix potential race conditions
The page load event, and page loading failed event can both be fired.
Consequently, there is a chance that we `select` the `pageReady`
event eventhough there may be a `pageFailed`.

To tackle this, we will pass the events as a struct through the same
channel. We can then wait on a single channel, and destructure it to
check for errors.
Additionally, the events are subscribed to prior to navigation.
The problem with this however, is that the loading failed event
relies on the request ID from navigation. Thus, we will have to
set the `requestID` _globally_, and if it is not defined, we will
simply check to ensure the type is `Document`. This is so we will
not accidentally fail the conversion if a non-primary (i.e. not
a `mainFrame`) resource fails to load.

The race conditions are more likely to occur in debug mode.
They rarely occur otherwise.

Finally, stack traces are attached to errors from external packages.