[Version 3] Add `window-status` plugin
This is similar to the `--window-status <windowStatus>` flag in `wkhtmltopdf`.
But unlike `wkhtmltopdf`, the `windowStatus` is not set as a flag.

When used, it will wait for `window.status` to be a string defined in
`WINDOW_STATUS` before rendering the page. `WINDOW_STATUS` is set
in the target HTML, for example:

```html
<script>
	var WINDOW_STATUS = "athenapdf";
</script>
```

If it is not set globally, the plugin will look for `ready` in
`window.status`.

Closes #113
Similar to #104