[Version 3] Add support for delaying the PDF rendering via JS plugin
This is similar to the `--javascript-delay <msec>` flag in
`wkhtmltopdf`.

When used, it will wait for `ATHENAPDF_DELAY` milliseconds
before rendering the page to PDF. `ATHENAPDF_DELAY` is set
in the target HTML, for example:

```html
<script>
    // Wait for 2s
    var ATHENAPDF_DELAY = 2000;
</script>
```

It can be used to give enough time for JavaScript on the page to
execute and/or for web assets, e.g. remote fonts, to load.

If it is not set globally, the plugin will wait for 1000ms / 1s.

This plugin can be used to resolve #117.