[Version 3] Create `Processor` interface for use with the manager
This commit removes the pre-assembled process in favour of
allowing generic processes that are 'assembled' by end-users.
They can be passed to the manager as long as they satisfy
the `Processor` interface. This is to improve extensibility
so that instrumentation can be added.

Also, `worker.go` has been renamed to `manager.go` for clarity
(it is an Athena 'process' manager).