Add timeout to CloudConvert job based on the worker timeout
`5` is an arbitrary number to reduce the likelihood of overlap.

It does not seem like the underlying job is cancelled when the
request is closed. Furthermore, the default timeout for pro
accounts is very high (5 hours). It is better if we timeout quickly
for documents that are large, and likely to fail.