Pass headers via weaver and store cookies in electron
This allows session emulation so this can be used as part of a script
to act as a logged-in user (for pages without CSRF protections).

The athenapdf file was mangled by my editor's default formatting settings,
but I'm putting this up to see if there any interest in this kind of feature.

I also made some quality of life improvements for those running this
with a single AWS account.