Fix `weaver` local conversion treating HTML files as plaintext
Due to a recent upgrade (f59e999), Chromium (via Electron) is not
automatically detecting HTML files by the contents. Instead, the
extension of the file must be set for it to be treated as (M)HTML.
Thus, this commit sets the default file extension to `html` unless
otherwise specified using the `ext` argument (see b1a16ac8). This
is so that files will not be treated as plaintext, and so we will
not end up converting HTML to a PDF showing the raw HTML, but rather
the rendered HTML.

For debugging purposes, the name of the temporary file now includes
`athena` for easy identification in the temporary directory.